Fix a bug where policies where ints and use a better trained l1 256 policy net Results of dev vs main (8+1, NULL, NULL, UHO_Lichess_4852_v1.epd): Elo: 69.05 +/- 26.25, nElo: 80.79 +/- 29.86 LOS: 100.00 %, DrawRatio: 39.23 %, PairsRatio: 2.04 Games: 520, Wins: 263, Losses: 161, Draws: 96, Points: 311.0 (59.81 %) Ptnml(0-2): [18, 34, 102, 40, 66], WL/DD Ratio: 8.27 LLR: 2.98 (-2.94, 2.94) [0.00, 10.00]

+++ src/search/search.cpp
@@ -143,7 +143,7 @@ void Node::deallocate() {
 }
 
 void Node::labelPolicies(float* raw) {
-    int policies[256];
+    float policies[256];
     float sum = 0.0;
 
     for (int i = 0; i < childCount; i++)
